Add podified-ci-testing-tcib repo
This DRLN tag would be used just for building the containers,
so the `podified-ci-testing` then will be only updated after
we ensure we can properly build the containers. The reason
for this is that we have a lot of jobs that expect to find
container tag for `podified-ci-testing` already in repository,
while the containers may never be pushed due to tcib issues.
The intermediate tag would allow us to bypass this without
redesigning all existing jobs.